+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.0] - 2022-08-16
+## [0.9.0] - 2022-08-17
 
 ### Added
 * Added the `Graph.AddVertex` method for adding a vertex. This replaces `Graph.Vertex`.